requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
google-cloud-storage
pydantic-settings 
cryptography